Introduced a Customer Class
For now, All i did was give it some attributes and create a basic LogIn method that will grab a file from the DB and fill out the customer class with the corresponding attributes. Unfortunately it will crash/fail if you put in wrong email and password so lot of works needs to be done. It is a prototype for LogIn/Grabbing customer file.